backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework_simplejwt